Started work on polygons
Started the code for polygons with an odd number of sides. Derived a
matrix equation to find the radii attached to each vertex and am
planning to add a document explaining the derivation later. For now the
code is in place, but I need to add routines for the actual graphing.